Refactor ValidationResult into MarkdownReader.
Isolate cmdlets from any specific serialization/deserialization.
