+++ src/Command/ImportYamlModuleFileCommand.cs
@@ -5,12 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Management.Automation;
-using System.Management.Automation.Runspaces;
-using System.Runtime.Serialization;
-using Microsoft.PowerShell.PlatyPS;
 using Microsoft.PowerShell.PlatyPS.Model;
-using YamlDotNet.Serialization;
-using YamlDotNet.Serialization.NamingConventions;
 
 namespace Microsoft.PowerShell.PlatyPS
 {
@@ -57,13 +52,13 @@ namespace Microsoft.PowerShell.PlatyPS
 
                 if (AsDictionary)
                 {
-                    if (YamlUtils.TryGetMetadataFromText(File.ReadAllText(path), out var dictionaryResult))
+                    if (YamlUtils.TryGetOrderedDictionaryFromText(File.ReadAllText(path), out var dictionaryResult))
                     {
                         WriteObject(dictionaryResult);
                     }
                     else
                     {
-                        WriteError(new ErrorRecord(new SerializationException("DeserializationError"), "ImportYamlModuleFile,FailedToConvertYamlToDictionary", ErrorCategory.InvalidOperation, path));
+                        WriteError(new ErrorRecord(new InvalidOperationException("DeserializationError"), "ImportYamlModuleFile,FailedToConvertYamlToDictionary", ErrorCategory.InvalidOperation, path));
                     }
                     continue;
                 }
