+++ src/Command/TestMarkdownCommandHelp.cs
@@ -2,41 +2,11 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.IO;
 using System.Management.Automation;
-using System.Text;
 
 namespace Microsoft.PowerShell.PlatyPS
 {
-    public class MarkdownCommandHelpValidationResult
-    {
-        public string Path { get; set; }
-        public bool IsValid { get; set; }
-        public List<string> Messages { get; set; }
-        public MarkdownCommandHelpValidationResult()
-        {
-            Path = string.Empty;
-            IsValid = true;
-            Messages = new List<string>();
-        }
-
-        public MarkdownCommandHelpValidationResult(string path, bool isValid, List<string> messages)
-        {
-            Path = path;
-            IsValid = isValid;
-            Messages = messages;
-        }
-
-        public override string ToString()
-        {
-            return $"Path: {Path}, IsValid: {IsValid}";
-        }
-
-    }
-
-
     [Cmdlet(VerbsDiagnostic.Test, "MarkdownCommandHelp")]
     [OutputType(typeof(bool))]
     [OutputType(typeof(MarkdownCommandHelpValidationResult))]
